# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/auctions/{auctionId}/bids" : {
      "get" : {
        "summary" : "경매의 입찰 내역을 금액 내림차순으로 페이징 조회합니다.",
        "description" : "경매의 입찰 내역을 금액 내림차순으로 페이징 조회합니다.",
        "parameters" : [ {
          "name" : "auctionId",
          "in" : "path",
          "description" : "경매 ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "page" : {
                    "type" : "integer",
                    "nullable" : false
                  },
                  "size" : {
                    "type" : "integer",
                    "nullable" : false
                  },
                  "sort" : {
                    "type" : "array",
                    "nullable" : false,
                    "items" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          },
          "required" : false
        },
        "responses" : {
          "200" : {
            "description" : "입찰 내역 목록 (페이징)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "content" : {
                      "type" : "array",
                      "description" : "현재 페이지의 데이터 목록",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "integer",
                            "description" : "입찰 ID",
                            "format" : "int64"
                          },
                          "auctionId" : {
                            "type" : "integer",
                            "description" : "소속 경매 ID",
                            "format" : "int64"
                          },
                          "bidderNickname" : {
                            "type" : "string",
                            "description" : "입찰자 닉네임"
                          },
                          "amount" : {
                            "type" : "number",
                            "description" : "입찰 금액"
                          },
                          "bidTime" : {
                            "type" : "string",
                            "description" : "입찰 시각"
                          },
                          "bidStatus" : {
                            "type" : "string",
                            "description" : "{ACTIVE=ACTIVE, OUTBID=OUTBID, WINNING=WINNING}",
                            "enum" : [ "ACTIVE", "OUTBID", "WINNING" ]
                          }
                        },
                        "description" : "입찰 응답 DTO."
                      }
                    },
                    "page" : {
                      "type" : "integer",
                      "description" : "현재 페이지 번호 (0-based)"
                    },
                    "size" : {
                      "type" : "integer",
                      "description" : "페이지 크기"
                    },
                    "totalElements" : {
                      "type" : "integer",
                      "description" : "전체 데이터 건수",
                      "format" : "int64"
                    },
                    "totalPages" : {
                      "type" : "integer",
                      "description" : "전체 페이지 수"
                    },
                    "first" : {
                      "type" : "boolean",
                      "description" : "첫 번째 페이지 여부"
                    },
                    "last" : {
                      "type" : "boolean",
                      "description" : "마지막 페이지 여부"
                    }
                  },
                  "description" : "입찰 내역 목록 (페이징)"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "입찰을 제출합니다.",
        "description" : "입찰을 제출합니다.",
        "parameters" : [ {
          "name" : "auctionId",
          "in" : "path",
          "description" : "경매 ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "nickname",
          "in" : "header",
          "description" : "입찰자 닉네임 (요청 헤더)",
          "required" : true,
          "schema" : { }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "bidAmount" ],
                "type" : "object",
                "properties" : {
                  "bidAmount" : {
                    "minimum" : 0,
                    "exclusiveMinimum" : false,
                    "type" : "number",
                    "description" : "입찰 금액. 현재가 + 최소 입찰 단위 이상이어야 합니다.\n                  첫 입찰의 경우 시작가 이상이어야 합니다."
                  }
                },
                "description" : "입찰 요청 데이터"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "입찰 결과 (HTTP 201)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "id" : {
                      "type" : "integer",
                      "description" : "입찰 ID",
                      "format" : "int64"
                    },
                    "auctionId" : {
                      "type" : "integer",
                      "description" : "소속 경매 ID",
                      "format" : "int64"
                    },
                    "bidderNickname" : {
                      "type" : "string",
                      "description" : "입찰자 닉네임"
                    },
                    "amount" : {
                      "type" : "number",
                      "description" : "입찰 금액"
                    },
                    "bidTime" : {
                      "type" : "string",
                      "description" : "입찰 시각"
                    },
                    "bidStatus" : {
                      "type" : "string",
                      "description" : "{ACTIVE=ACTIVE, OUTBID=OUTBID, WINNING=WINNING}",
                      "enum" : [ "ACTIVE", "OUTBID", "WINNING" ]
                    }
                  },
                  "description" : "입찰 결과 (HTTP 201)"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}